requirements.txt
fastapi
uvicorn
python-dotenv
openai
pdfplumber
pandas
python-multipart